🗃Refactor: 기능 함수 분리
