+++ src/js/@helper/index.js
@@ -1,3 +1,4 @@
 export * from './dom/index.js';
+export * from './publisher.js';
 export * from './util.js';
 export * from './valid.js';
